fazendo scaffold

+++ app/src/main/java/br/senai/sp/jandira/login_screen/model/Categories.kt
@@ -3,7 +3,7 @@ package br.senai.sp.jandira.login_screen.model
 import androidx.compose.ui.graphics.painter.Painter
 
 data class Categories
-    (var id: Int = 0,
+    (var id: Long = 0,
     var name: String = "",
     var image: Painter? = null   )
 {}
